definitely ready for deployment 🤠

+++ AskMate/Repos/Votes/IVotesRepo.cs
@@ -11,6 +11,6 @@ namespace AskMate.Repos.Votes
         public void ChangeVote(int voteId);
         public void DeleteVote(int voteId);
 
-        float CalculateUserAnswersUsefulnessRatio(string userId);
+        decimal CalculateUserAnswersUsefulnessRatio(string userId);
     }
 }
